Update rename & rearrange passport.js
