+++ server/routes/userFriend.js
@@ -2,12 +2,12 @@
  * @Author: Hu Keyi
  * @Date: 2021-05-06 20:20:16
  * @Last Modified by: Hu Keyi
- * @Last Modified time: 2021-05-22 13:01:16
+ * @Last Modified time: 2023-03-06 10:57:38
  */
 const express = require('express');
 const router = express.Router();
 const friendController = require('../controllers/userFriend.js');
-const passport = require('../configs/passport.config.js');
+const passport = require('../utils/passport.js');
 
 router.get(
 	'/list',
