+++ server/utils/passport.js
@@ -2,7 +2,7 @@
  * @Author: Hu Keyi
  * @Date: 2021-05-04 22:46:28
  * @Last Modified by: Hu Keyi
- * @Last Modified time: 2021-06-02 13:37:36
+ * @Last Modified time: 2023-03-06 10:56:25
  */
 
 // passport setting
@@ -22,11 +22,11 @@ options.secretOrKey = process.env.PASSPORT_JWT_SECRET;
 
 // 序列化与反序列化
 passport.serializeUser((user, done) => {
-	console.log('\n-----> serialize user', toJSON(user));
+	console.log('\n【Passport.js】serialize user', toJSON(user));
 	done(null, user.id);
 });
 passport.deserializeUser(async (id, done) => {
-	console.log('\n-----> deserialize user', id);
+	console.log('\n【Passport.js】deserialize user', id);
 	const user = await User.findOne({
 		where: {
 			id: id,
@@ -38,9 +38,9 @@ passport.deserializeUser(async (id, done) => {
 // 在需要拦截未验证的用户的请求的时候调用
 passport.ensureAthenticated = function ensureAthenticated() {
 	return function (req, res, next) {
-		console.log('\n-----> in ensureSthenticated process:');
+		console.log('\n【Passport.js】in ensureSthenticated process:');
 		if (req.isAuthenticated()) {
-			console.log('\n-----> Permission allowed');
+			console.log('\n【Passport.js】Permission allowed');
 			//这里是passport在登录验证时已经设置好的加密cookie
 			return next();
 		}
@@ -76,23 +76,31 @@ passport.use(
 						],
 					})
 				);
-				console.log('\n-----> user ' + username + ' attempted to log in.');
+				console.log(
+					'\n【Passport.js】user ' +
+						username +
+						' attempted to log in.'
+				);
 
 				// 用户不存在
 				if (!user) {
-					console.log('\n-----> user not existed');
+					console.log('\n【Passport.js】user not existed');
 					return done(null, false);
 				}
 				// 密码不正确
 				if (!bcrypt.compareSync(password, user.password)) {
-					console.log('\n-----> wrong password');
+					console.log('\n【Passport.js】wrong password');
 					return done(null, false);
 				}
-				console.log('\n-----> user info from database:', user.id, user.name);
+				console.log(
+					'\n【Passport.js】user info from database:',
+					user.id,
+					user.name
+				);
 				delete user.password;
 				return done(null, user);
 			} catch (err) {
-				console.log('\n-----> something went wrong', err);
+				console.log('\n【Passport.js】something went wrong', err);
 				return done(err);
 			}
 		}
@@ -105,16 +113,20 @@ passport.use(
 		// fixme: 后面做邮箱验证的话，可能需要加一个字段判断是手机号还是邮箱
 		try {
 			const user = await User.findOne({ where: { id: jwt_payload.id } });
-			console.log('\n----->jwt got userinfo from db:', user.id, user.name);
+			console.log(
+				'\n【Passport.js】jwt got userinfo from db:',
+				user.id,
+				user.name
+			);
 
 			// 用户不存在
 			if (!user) {
-				console.log('\n-----> user not existed');
+				console.log('\n【Passport.js】user not existed');
 				return done(null, false);
 			}
 			return done(null, user);
 		} catch (err) {
-			console.log('\n-----> jwtstrategy went wrong', err);
+			console.log('\n【Passport.js】jwtstrategy went wrong', err);
 			return done(err);
 		}
 	})
